+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Show maximum mana in Wand lore.
  - "damage_reduction" properties changed to "protection".
+ - Add boon + curse spells (PotionEffectSpell)
 
 ## 2.5.9